+++ lib/data/dummy_data.dart
@@ -1,89 +1,137 @@
 class DummyData {
   static final events = [
     <String, dynamic>{
-      'id': 'test1',
-
-      'name': 'Concert',
+      'id': 'test1a',
+      'name': 'Orange Warsaw Festival',
       'description':
           'Orange Warsaw Festival 2022 is coming and announcing a big part of it’s line-up. Florence + The Machine and Tyler, The Creator headline one of the biggest music festivals in Poland. Also joining the list are Charli XCX, NAS, Foals, Stormzy, Earl Sweatshirt and Sigrid, as well as great local acts - Żabson x Young Igi, Szczyl and Kacperczyk. See you at Horse Racetrack Warszawa - Służewiec, June 3-4.',
       'place': {'lat': 52.1597, 'lon': 21.0087},
       'images': [
-        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg'
+        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg',
+        'https://orangewarsawfestival.pl/uploads/media/x500/00/990-OWF2019_ZSosnowska_Solange_13_lowres.jpg?v=1-0',
+        'https://www.glamour.pl/media/cache/default_view/uploads/media/default/0006/22/f4c7d0500c517f260619facd77f62e9f0738a778.jpg'
       ]
     },
     <String, dynamic>{
-      'id': 'test2',
-
+      'id': 'test2a',
       'name': 'Silent Disco',
       'description':
           'No plans for Saturday ?! An amazing event in Warsaw, Juwenalia University of Music and the quietest party in the capital, Silent Disco. Follow closely!',
       'place': {'lat': 53.1096, 'lon': 23.1475},
       'images': [
-        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg'
+        'https://cdn.getyourguide.com/img/tour/5cf4f4d8ca3fa.jpeg/148.jpg',
+        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRp7oqYZFF6tD-jiWzwz5X8NkDSHyoUAfAzDQ&usqp=CAU',
+        'https://goingapp.pl/storage/app/uploads/public/5d2/707/004/thumb_408731_600x350_0_0_crop.jpg'
       ]
     },
     <String, dynamic>{
-      'id': 'test3',
-
+      'id': 'test3a',
       'name': 'Outdoor cinema',
       'description':
           'Powiśle power plant together with Millennium Docs Against Gravity invite you to the unique THIRD edition of summer cinema in Powiśle!' +
               'For the coming months, we meet every Tuesday and Saturday at 19:30 at the Power Plant.',
       'place': {'lat': 52.2414, 'lon': 21.0275},
       'images': [
-        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg'
+        'https://mldvwwasb8tu.i.optimole.com/5CC9Lzs.esbD~6200b/w:1700/h:1133/q:90/https://travelaway.me/wp-content/uploads/2016/05/kamari-cinema.jpg',
+        'https://media.timeout.com/images/105766007/image.jpg',
       ]
     },
-    <String, dynamic>{
-      'id': 'test4',
-      'name': '',
-      'description': '',
-      'place': {'lat': 0.0, 'lon': 0.0},
-      'images': [
-        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg'
-      ]
-    }
+
   ];
   static final places = [
     <String, dynamic>{
-      'id': 'test1',
+      'id': 'test1b',
       'name': 'Łazienki Królewskie',
       'description': 'The Royal Łazienki Museum, called "the happiest place in Warsaw", was the summer residence of King Stanisław August. It is made up of classicist monuments and historical gardens.',
       'place': {'lat': 52.2156, 'lon': 21.0349},
       'images': [
-        'https://warsawtour.pl/wp-content/uploads/2018/08/%C5%81azienki-Kr%C3%B3lewskie-Pa%C5%82ac-na-Wyspie_fot.-z-archiwum-Warszawskiej-Organizacji-Turystycznej.jpg'
+        'https://warsawtour.pl/wp-content/uploads/2018/08/%C5%81azienki-Kr%C3%B3lewskie-Pa%C5%82ac-na-Wyspie_fot.-z-archiwum-Warszawskiej-Organizacji-Turystycznej.jpg',
+        'https://upload.wikimedia.org/wikipedia/commons/f/fb/Pa%C5%82ac_na_wodzie_w_%C5%81azienkach_Kr%C3%B3lewskich_%28cropped%29.jpg',
+        'https://pliki.propertydesign.pl/i/02/13/95/021395_r0_1140.jpg'
       ],
-      'video': 'https://www.youtube.com/watch?v=0QnPuZ3AfPc'
+      'video': '0QnPuZ3AfPc'
     },
     <String, dynamic>{
-      'id': 'test2',
+      'id': 'test2b',
       'name': 'Royal Castle in Warsaw',
       'description': 'With its characteristic shape and distinct colour, you will not mistake this building for any other. The Royal Castle is the former seat of the rulers of Poland. Go inside and look at the royal apartments and the Throne Room. Pay attention to original paintings by Rembrandt and the canvases of Bernardo Bellotto, known as Canaletto, which were used during the reconstruction of Warsaw after it was destroyed in World War II. See also the Senator’s Hall, in which in 1791 the 3rd May Constitution was passed – the first constitution in Europe and the second in the world. At the end, enter the beautifully reconstructed gardens: the upper garden situated on the castle slope and the lower one at its feet where the river used to flow in the seventeenth century. Carefully recreated in the style of the 1920s and 1930s, it not only complements the residence but it is also an ideal place to relax. The huge lawn, flower beds, fountains and mazes created from several-metre high hedges will give you a boost of energy for further sightseeing. If you have more time, stay a bit longer to see the charm of the garden in the glow of the illuminations after dark.',
       'place': {'lat': 52.2481, 'lon': 21.0152},
       'images': [
-        'https://warsawtour.pl/wp-content/uploads/2018/11/Zamek-Kr%C3%B3lewski-fot.-Roman-Gorielov_Fotolia.jpg'
+        'https://warsawtour.pl/wp-content/uploads/2018/11/Zamek-Kr%C3%B3lewski-fot.-Roman-Gorielov_Fotolia.jpg',
+      'https://ocdn.eu/pulscms-transforms/1/k2Lk9kqTURBXy9jNDBjNTI1MjlhNzg1NTExNGJlNTcwOTE5YTQ0MzYwYS5qcGVnk5UDAD3NC7jNBpeTCaYzNTIzYWIGkwXNBLDNAnaBoTAB/zamek-krolewski-warszawa.jpg',
+      'https://sds.jedwabno.pl/wp-content/uploads/2020/11/DJI_0247-2-930x620.jpg'
       ],
-      'video': 'https://www.youtube.com/watch?v=0QnPuZ3AfPc'
-    }
+      'video': 'A2n1Jk4rLHA'
+    },
+    <String, dynamic>{
+      'id': 'test3b',
+      'name': 'Pałac Kultury i Nauki',
+      'description': "On 2000/2001 New Year's Eve, a clock was kicked off at the top of PKiN – it is the second biggest clock in Europe and has four dials, each 6 meters in diameter.",
+      'place': {'lat': 52.2317, 'lon': 21.0057},
+      'images': [
+        'https://i.wpimg.pl/O/644x467/d.wpimg.pl/1982839758--114368225/Palac_Kultury_i_Nauki.jpg',
+        'https://dzieje.pl/sites/default/files/styles/open_article_750x0_/public/201807/pkin.jpg?itok=GfNUdNRO',
+        'http://doc.rmf.pl/rmf_fm/sd/galeria/main/mi/palace-of-culture-and-science-441262_1920.jpg'
+      ],
+      'video': 'OEJ7f46o7tM'
+    },
+<String, dynamic>{
+      'id': 'test4b',
+      'name': 'Obiekto - bouldering',
+      'description': "The facility is located at ul. Mińska 65 (entrance from ul. Chodakowska, through the barrier opposite the ORLEN station). You will find plenty of light, atmospheric space and climbing for both beginners and advanced. On our wall you will start your climbing adventure from scratch, develop your skills and climb with the best climbers in the country.",
+      'place': {'lat': 52.2513, 'lon': 21.0670},
+      'images': [
+        'https://www.vogue.pl/uploads/repository/miejsca_na_powrot_do_formy/obiekto.jpg',
+        'https://img.grouponcdn.com/deal/3rt3b34tdvu2TS9ygJYibUr8562n/3r-2048x1229/v1/t600x362.jpg'
+      ],
+      'video': 'UXx4eKyPwUQ'
+    },
+    <String, dynamic>{
+      'id': 'test5b',
+      'name': 'Copernicus Science Centre',
+      'description': "We invite all Ukrainian families to join our cosmic adventure. In the Copernicus Science Centre Planetarium, we will organise free screenings of our two acclaimed titles – Polaris (age 5+) and We Are Aliens (age 10+). The screenings will be preceded with live sky shows, also in Ukrainian.",
+      'place': {'lat': 52.2420, 'lon': 21.028},
+      'images': [
+        'https://t0.gstatic.com/images?q=tbn:ANd9GcS8aBznZQW951Y_F2T4Feshk2DFFVouB7jJtJP4eLE6XwRivbGB',
+        'https://poland.pl/media/public/3b/19/5fee97d41a3acc2f87d6fab2408.jpg__768x600_q85_crop-smart_subsampling-2_upscale.jpg',
+        'https://warsawtour.pl/wp-content/uploads/2018/08/Copernicus-Science-Centre-4_fot.-Centrum-Nauki-Kopernik.jpg'
+      ],
+      'video': '1dcAOR9PCmg'
+    },
   ];
   static final accomodation = [
     <String, dynamic>{
-      'id': 'test1',
+      'id': 'test1c',
       'name': 'Mercure Warszawa',
       'description': 'Offering a bar, Mercure Warszawa Ursus Station is located in Warsaw in the Masovian Region Region. The distance to important places from the hotel is: Blue City Shopping Center - 6 km, Warsaw Uprising Museum - 8 km. Facilities include a restaurant, a 24-hour front desk, room service and free WiFi in all areas. The hotel offers a sauna and luggage storage.',
       'place': {'lat': 52.2156, 'lon': 21.0349},
       'images': [
-        'https://www.konferencje.pl/gfx/konferencje/userfiles/images/Event%20czym%20jest%20event%20w%20plenerze.jpg'
+        'https://www.ahstatic.com/photos/3384_ho_00_p_1024x768.jpg',
+        'https://images.sembo.se/ImageService/ImageHandler.ashx?service=sembo&nameOfImage=87769.jpg&resizeMode=FitOutside&width=1220&height=768&formatSettings=jpeg(quality-90)',
+        'https://foto.hrsstatic.com/fotos/0/3/545/350/80/000000/http%3A%2F%2Ffoto-origin.hrsstatic.com%2Ffoto%2F0%2F1%2F1%2F7%2F011733%2F011733_a_27312725.jpg/vCqI9Bhc4H8UA3sIAu1ygA%3D%3D/2048,1536/6/Hotel_Mercure_Warszawa_Centrum-Warschau-Aussenansicht-11-11733.jpg'
+      ],
+    },
+    <String, dynamic>{
+      'id': 'test2c',
+      'name': 'HOTEL WARSZAWA',
+      'description': 'One of the most important representatives of Polish interwar architecture. It became an important the symbol of Warsaw revival after the war. Towering majestically above the capital, visible from a distance of up to 20 km. Its modern steel and reinforced concrete construction was the answer to the strong and dynamic urbanization of the Warsaw agglomeration and proof of the mastery of pre-war engineering. Today, it is one of the most modern hotels in Poland.',
+      'place': {'lat': 52.2358, 'lon': 21.0126},
+      'images': [
+        'https://manager.lbooking.online/CmsImages/Hotel-Warszawa-pieciogwiazdkowy-hotel-w-centrum-warszawy-1(3).jpg'
+        ,'https://manager.lbooking.online/CmsImages/Hotel-Warszawa-pieciogwiazdkowy-hotel-w-centrum-warszawy-24(1).jpg',
+        'https://manager.lbooking.online/CmsImages/Hotel-Warszawa-pieciogwiazdkowy-hotel-w-centrum-warszawy-36(1).jpg'
+      ],
+    },
+    <String, dynamic>{
+      'id': 'test3c',
+      'name': 'Leonardo Royal Hotel',
+      'description': 'A hotel located in a skyscraper with bright rooms and apartments, an elegant restaurant and a simple bar.',
+      'place': {'lat': 52.2344, 'lon': 20.9901},
+      'images': [
+        'https://static.leonardo-hotels.com/image/E_LRHW_03_APR17_4000x2600_1200x780_mobile_3.webp'
+        ,'https://static.leonardo-hotels.com/image/E_LRHW_03_JUN17_4000x2600_1200x780_mobile_3.webp',
+        'https://static.leonardo-hotels.com/image/RCDBL_LRHW_01_FEB18_4000x2600_1200x780_mobile_3.webp'
       ],
     }
-  ];
-
-  static final toures = [
-    <String, dynamic>{
-      'id': 'test1',
-      'name': 'Guide 1',
-      'descrption': 'First warsaw guide',
-      'places': ['test1', 'test2']
-    },
   ];
 }