+++ lib/generated_plugin_registrant.dart
@@ -8,6 +8,7 @@
 
 import 'package:firebase_auth_web/firebase_auth_web.dart';
 import 'package:firebase_core_web/firebase_core_web.dart';
+import 'package:flutter_tts/flutter_tts_web.dart';
 import 'package:geolocator_web/geolocator_web.dart';
 import 'package:location_web/location_web.dart';
 import 'package:shared_preferences_web/shared_preferences_web.dart';
@@ -18,6 +19,7 @@ import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 void registerPlugins(Registrar registrar) {
   FirebaseAuthWeb.registerWith(registrar);
   FirebaseCoreWeb.registerWith(registrar);
+  FlutterTtsPlugin.registerWith(registrar);
   GeolocatorPlugin.registerWith(registrar);
   LocationWebPlugin.registerWith(registrar);
   SharedPreferencesPlugin.registerWith(registrar);