+++ lib/models/base_info.dart
@@ -1,3 +1,4 @@
+import 'package:city_guide_app/utils/utils.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
@@ -5,6 +6,8 @@ import 'package:uuid/uuid.dart';
 import 'package:json_annotation/json_annotation.dart';
 
 @immutable
+@JsonSerializable()
+@LatLngConverter()
 abstract class BaseInfo {
   final String id;
   final String name;
@@ -14,6 +17,8 @@ abstract class BaseInfo {
 
   const BaseInfo(this.name, this.description, this.place, this.images,
       {required this.id});
+
+  
 
   Icon get typeIcon => Icon(Icons.abc);
 }