+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 28-05-2022
+
+### Added:
+* Describtion tts
+* Added yt videos 
+* Added guided tour screen
+* Added guided tour info
+* Added guide card component
+* Added dummy data
+* Guide provider
+
+### Changed:
+* item card network image to network cached image
 
 ## [0.2.0] - 26-05-2022